refactor: extracts keyboard action logic into a custom hook
Refactors the Scene component to improve code organization
by extracting keyboard event handling and related logic into
a reusable `useKeyboardActions` custom hook.

This change enhances maintainability and readability of the
Scene component, centralizing keyboard-related functionality
in a dedicated module. Also adds a helper function to get item
dimensions.
