+++ src/components/Builder/Scene.js
@@ -21,33 +21,21 @@ import {
   findItemById,
   findItemsOnPage,
   getItemsInSelectionBox,
-  getMostVisiblePage,
 } from '../../utils/functions';
-import { useSelectedElements, useEventListener } from '../../utils/hooks';
-import generateId from '../../utils/generateId';
-import { EVENT_IGNORED_ROLES } from '../../constants/eventIgnoredRoles';
 import DraggableLayer from './DraggableLayer';
+import useKeyboardActions from '../../utils/useKeyboardActions';
 
 const Scene = () => {
   const pages = usePropStore(state => state.pages);
-  const onItemAdd = usePropStore(state => state.onItemAdd);
-  const onItemRemove = usePropStore(state => state.onItemRemove);
-  const onAnEventTrigger = usePropStore(state => state.onAnEventTrigger);
   const settings = usePropStore(state => state.settings);
-  const onItemChange = usePropStore(state => state.onItemChange);
-
-  const activeElements = useBuilderStore(state => state.activeElements);
   const contextMenuProps = useBuilderStore(state => state.contextMenuProps);
-  const isRightPanelOpen = useBuilderStore(state => state.isRightPanelOpen);
-  const setActiveElements = useBuilderStore(state => state.setActiveElements);
+
   const resetActiveElements = useBuilderStore(state => state.resetActiveElements);
   const setActiveElementsSelection = useBuilderStore(state => state.setActiveElementsSelection);
   const setContextMenuProps = useBuilderStore(state => state.setContextMenuProps);
   const setGuides = useBuilderStore(state => state.setGuides);
-  const setIsRightPanelOpen = useBuilderStore(state => state.setIsRightPanelOpen);
-  const zoom = useBuilderStore(state => state.zoom);
 
-  const [itemToPaste, setItemToPaste] = useState(null);
+  const zoom = useBuilderStore(state => state.zoom);
 
   // Marquee selection state
   const [isSelecting, setIsSelecting] = useState(false);
@@ -66,6 +54,9 @@ const Scene = () => {
     return acc;
   }, {}));
 
+  // Handle keyboard actions
+  useKeyboardActions();
+
   // Update refs when new pages are added
   useEffect(() => {
     pages.forEach(page => {
@@ -74,11 +65,6 @@ const Scene = () => {
       }
     });
   }, [pages]);
-
-  const isMultipleItemSelected = activeElements.length > 1;
-
-  /* Calculate snap guides */
-  const keyDownCount = useRef(null);
 
   useEffect(() => {
     setGuides(pages.reduce((acc, page) => {
@@ -112,180 +98,6 @@ const Scene = () => {
       viewPortRef.current.scrollTop = lastScrollPosition;
     }
   }, [lastScrollPosition]); // set last scroll position after changing mode
-
-  const foundItem = findItemById(activeElements.length ? activeElements[0] : null, pages);
-
-  const selectedItems = useSelectedElements();
-  const moveItemWithKeyboard = (event, direction, value) => {
-    event.preventDefault();
-    selectedItems.forEach(item => {
-      if (item.isLocked) {
-        return false;
-      }
-      onItemChange(
-        { id: item.id },
-        {
-          ...item,
-          [direction]: item[direction] + value,
-        },
-      );
-    });
-  };
-
-  const selectNextOrPrevElement = (event, deletedItem) => {
-    if (event.preventDefault) event.preventDefault();
-    const referenceItem = deletedItem ? deletedItem : foundItem;
-
-    const page = pages.find(_page => _page.id === referenceItem.pageID);
-    if (!page || (page && !page.items.length)) return resetActiveElements();
-    const { items } = page;
-
-    const currentIndex = items.findIndex(item => item.id === referenceItem.id);
-
-    // Pages are not updated in time so here is an unnecessary check
-    if (items.length === 1 && deletedItem) {
-      return resetActiveElements();
-    }
-
-    if (event.shiftKey) {
-      if (items[currentIndex - 1]) setActiveElements(items[currentIndex - 1].id);
-      else setActiveElements(items[items.length - 1].id);
-    } else if (items[currentIndex + 1]) {
-      setActiveElements(items[currentIndex + 1].id);
-    } else setActiveElements(items[0].id);
-  };
-
-  const onItemRemoveFromPage = e => {
-    // Firefox updates browser history on backspace
-    e.preventDefault();
-    if (isMultipleItemSelected) return;
-    if (foundItem.isLocked) {
-      return false;
-    }
-    resetActiveElements();
-    onItemRemove(foundItem);
-    selectNextOrPrevElement({ shiftKey: false }, foundItem);
-    onAnEventTrigger('removeItem', foundItem.itemType);
-  };
-
-  const handlePaste = () => {
-    if (isMultipleItemSelected) return;
-    const itemID = generateId();
-    const pageID = getMostVisiblePage(true);
-    const offset = itemToPaste.pageID === pageID ? 50 : 0;
-
-    const item = {
-      ...itemToPaste,
-      id: itemID,
-      left: itemToPaste.left + offset,
-      pageID,
-      top: itemToPaste.top + offset,
-    };
-
-    onItemAdd(item);
-
-    onAnEventTrigger('pasteItem', itemToPaste.itemType);
-    setActiveElements(itemID);
-    // set as last reference to paste
-    setItemToPaste(item);
-  };
-
-  const keyboardActions = event => {
-    const {
-      key,
-      metaKey,
-      shiftKey,
-    } = event;
-
-    if (metaKey) {
-      if (key === 'l') {
-        // Lock
-        if (isMultipleItemSelected) return;
-        event.preventDefault(); // Dont focus to URL bar
-        onAnEventTrigger(
-          foundItem.isLocked ? 'unlockReportItem' : 'lockReportItem',
-          foundItem.itemType,
-        );
-        onItemChange(
-          { id: foundItem.id },
-          { isLocked: foundItem.isLocked ? false : true },
-        );
-        return;
-      }
-
-      if (key === 'c' || key === 'x') {
-        if (isMultipleItemSelected) return;
-        // Copy or Cut
-        if (key === 'x') {
-          onItemRemoveFromPage(event);
-          onAnEventTrigger('cutItem', foundItem.itemType);
-        } else {
-          onAnEventTrigger('copyItem', foundItem.itemType);
-        }
-
-        setItemToPaste(foundItem);
-        return;
-      }
-
-      if (foundItem && key === 'd') {
-        if (isMultipleItemSelected) return;
-        // Duplicate
-        event.preventDefault();
-        const itemID = generateId();
-        onItemAdd({
-          ...foundItem,
-          id: itemID,
-          left: foundItem.left + 50,
-          top: foundItem.top + 50,
-        });
-        onAnEventTrigger('duplicateItem', foundItem.itemType);
-        setActiveElements(itemID);
-        if (!isRightPanelOpen) {
-          setIsRightPanelOpen(true);
-        }
-        return;
-      }
-    }
-
-    keyDownCount.current++;
-    const movementValue = shiftKey ? 10 : 1 + keyDownCount.current;
-
-    // Others
-    switch (key) {
-    case 'Backspace': return onItemRemoveFromPage(event);
-    case 'Delete': return onItemRemoveFromPage(event);
-    case 'Escape': return resetActiveElements();
-    case 'ArrowLeft': return moveItemWithKeyboard(event, 'left', -movementValue);
-    case 'ArrowUp': return moveItemWithKeyboard(event, 'top', -movementValue);
-    case 'ArrowRight': return moveItemWithKeyboard(event, 'left', movementValue);
-    case 'ArrowDown': return moveItemWithKeyboard(event, 'top', movementValue);
-    case 'Tab': return selectNextOrPrevElement(event);
-    default:
-    }
-  };
-
-  const shouldSuppressKeyboardEvent = e => (
-    EVENT_IGNORED_ROLES.some(role => e.target.closest(`[role=${role}]`))
-  );
-
-  const handleKeyboardEvent = e => {
-    const shouldPaste = itemToPaste && e.key === 'v' && e.metaKey;
-
-    if (shouldSuppressKeyboardEvent(e)) {
-      return;
-    }
-
-    if (activeElements.length && !shouldPaste) {
-      const arrowKeyCodes = ['ArrowLeft', 'ArrowUp', 'ArrowDown', 'ArrowBottom'];
-      if (arrowKeyCodes.includes(e.key)) e.preventDefault();
-      keyboardActions(e);
-    } else if (shouldPaste) {
-      handlePaste();
-    }
-  };
-
-  useEventListener('keydown', handleKeyboardEvent);
-  useEventListener('keyup', () => { keyDownCount.current = 0; });
 
   // Marquee selection handlers
   const handleCanvasMouseDown = useCallback(e => {
